Animals
Lots of rework. File now just reads and writes to the file. Animal utils does the work of taking user input and adding a new animal to the list.

The program adds 2 extra entries of the last animal when a new one is added - need to find  the fix for this.